+++ src/Board.h
@@ -12,20 +12,20 @@ class Board
 public:
 
     Board( BoardSize, BoardMode mode );
-    Board( std::vector< uint >&& values, BoardSize, BoardMode mode );
+    Board( std::vector<size_t>&& values, BoardSize, BoardMode mode );
     Board( const Board& ) = delete;
     Board( Board&& ) = delete;
     Board& operator=( const Board& ) = delete;
     Board& operator=( Board&& ) = delete;
 
-    Move checkMove( uint row, uint col );
+    Move checkMove( size_t row, size_t col );
 
     BoardSize getSize() const;
     uint getSizeInt() const;
     BoardMode getMode() const;
     void setMode( BoardMode );
 
-    std::vector<uint>& getBoardValues();
+    std::vector<size_t>& getBoardValues();
     void generateBoard();
     void solveBoard();
 
@@ -33,16 +33,16 @@ public:
 
 private:
 
-    void makeMove( uint srcRow, uint srcColumn, uint dstRow, uint dstColumn );
+    void makeMove( size_t srcRow, size_t srcColumn, size_t dstRow, size_t dstColumn );
     uint findEmptyTilePosition();
 
     const BoardSize size;
-    const uint sizeInt;
+    const size_t sizeInt;
 
     BoardMode mode;
-    const uint nullValue;
+    const size_t nullValue;
 
-    std::vector<uint> values;
+    std::vector<size_t> values;
 };
 
 #endif
